Added Anime Class and search/getAnime functions
Added an Anime Class which is used to wrap the anime data return from the Kitsu API by the added searchAnime & getAnime functions.

The Anime Class is used to wrap around the data returned by the Kitsu API to give the user a consistent set of data to work with compared to Kitsus default. searchAnime allows the user to pass some search arguments as well as a page offset to the function to return an array of 10 Anime Objects which were found with the search. The page offset is used to return the anime from a specific page and by default is 0. getAnime allows the user to pass an anime ID which they would like to return an Anime Object for.