Style changes and slight fixes
Previously there were a couple bits of code that werent code the best, this has been fixed.

Instead of creating a new variable for the formatted anime, it is now directly passed the the promise resolve. pageOffset is no longer converted to string to check if it should be 0 or the value be kept, this was not needed. A few style changes with the positioning of code was also fixed.