Added Manga Class and search/getManaga functions
The Manga Class is used to wrap the data returned from the api by the searchManga and getManga functions.

The searchManga function allows you to Kitsu's database of manga based on the passed search args and can also take a page offset to return manga for that specific page. getManga takes a manga id and returns a manga object for that id if one exists.