Added User Class and search/getUser function
Added a User Class to wrap the data returned from the searchUser and getUserFunctions

The searchUser function allows you to pass a username to search for as well as the number of users to return, defaulted to 10, and will return an array of User Objects. The getUser function allows you to pass a user ID to return a User Object matching that ID. The User Class wraps the data returned from the Kitsu API to give a consistent set of data to work with.