Added Character Class and search/getCharacter.
Added the ability to search for or get a characters class data.

searchCharacters takes search args and a pageOffset and returns a character class.
getCharacter takes a character id and returns a character class.
Character Class wraps the data from the Kitsu API into a nice consistent package for ease of use.