Added missing Manga Class Properties
Previously many properties where missing from the manga class.

Added the volumeCount/chapterCount which can be used to see how long a manga is.
Added the ageRating/ageRatingGuide so the user is able to tell the rating of a manga.
Added ratingRank/popularityRank which are used for seeing how a manga is ranked.